medsearch-ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
endee
sentence-transformers==3.4.1
google-generativeai==0.8.5
python-dotenv==1.0.1
httpx>=0.28.1
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.4.0
pydantic==2.7.4
aiofiles==23.2.1
